requirements.txt
streamlit
openai
psycopg2-binary
pgvector
python-dotenv
pypdf
faiss-cpu
numpy
